+++ bin/js/utils/index.ts
@@ -2,17 +2,16 @@ const { warning, error, success } = require('./colors')
 const { writeFile, isFileExist } = require('./file')
 const { isCacheExpired, updateCacheTime } = require('./cache')
 const { createArgs } = require('./command')
+const getSeconds = require('./getSeconds')
 
 module.exports = {
     warning,
     error,
     success,
-
     writeFile,
     isFileExist,
-
     isCacheExpired,
     updateCacheTime,
-
-    createArgs
+    createArgs,
+    getSeconds
 }